extract RenderConfig
